src/requirements.txt
pycryptodome
loguru
python-dotenv
mysql-connector-python